"Dynamic" linearity enforcement
We’re now enforcing linearity the same way we did in cavy-python: during
compilation, we interpret the program line-by-line, and implement
quantum instructions by pushing gates onto a circuit. When variables
referring to quantum values are used, those variables are removed from
the environment. This isn’t proper static enforcement of linearity, but
I *think* it’s equivalent. It’s also apparently easier, as it doesn’t
require another separate compiler pass. However, it may turn out to be
less robust and more painful in the long run: cf. "shotgun parsing."

The more dynamic approach also nicely enables the repl; this would be a
little harder with a truly statically-typed version of the language.

In summary, this might be ripped out in the future and replaced with a
separate compiler pass. This shouldn’t be hard to do, as the changes are
confined to the `Environment` and the `is_linear` method now on the
`Value` type.