Array values, Cavy types
Arrays have been added to the Value enum. Because of the need to check
their type (e.g. their linearity) even then they are empty, my hand has
been forced to include a type field in this enum variant, and therefore
also a Type type. On one hand this is a large addition for such a small
aspect of one language feature, but on the other hand it should ease the
transition to purely static type checking in a future version of the
language.