Interpretation of array expressions
We’ve had to make a slightly awkward choice about the dynamic type of
arrays. Because values--not expressions--are typed, we need to make a
choice for the unique type of empty arrays. We’ve gone with the choice
that the type field of `T_Array` is an `Option<Type>`, which takes the
value `None` on an empty array. The invariant that this is `None` if and
only if the length field equals `0` cannot, as far as I'm aware, be
enforced statically in Rust.