Let expression statements
`let` _expressions proper_ are still not allowed; they don't pass the
parser. The reason is that we actually need a _borrow checker_ to admit
them. Absent any sort of static analysis, all else being equal, the code

```
let x = ?false;
let y = (let z = x in {
    z
});
```

is correct, yet clones a qubit. So, in the current state of the language,
we can't allow `let` blocks that return anything!