Spans, Ident nodes, SrcStore
* Locations are now tracked with a `Span` struct, similar to rustc.
Each `Span` contains an Id for some "source object", which can be resolved by looking it up in an appropriate source store.
* Added an `Ident` node to the AST.
This simplifies a whole lot of matches throughout the compiler.
* SrcStore
There's now a SrcStore type containing a map from `SrcId`s to source objects. It's not clear if this is really the right way to do this: it creates some new, 'quasi-global' state. But it does make it possible to track where errors are coming from, even with multiple source files, or mixed source files and repl input.