Error diagnostic macro;
Shamelessly cribbing from rustc yet again, I’ve added a proc
macro (`#[derive(Diagnostic)]`) for building error types. Like the
analogous macro in rustc, a particular member field is annotated with
the error message. This field should be a `Span`, which will be
pretty-printed along with the formatted error. The macro itself might be
a little inefficient, but I expect this not to matter very much,
especially given that the cost is paid only at compile-time.