Refactor: Session struct, unified error handling
The "global" data associated with a compilation process is now captured in a
`Session` struct. Each phase now has a single entry-point function,
which receives an reference to this struct and uses it to handle errors
if necessary. This has two advantages: first, it unifies the disparate
structures used to pass around "globals"; second, it made it easier to
propagate errors. Previously, there were multiple mechanisms, with some
passes maintaining a struct-local error buffer, and others propagating
(and accumulating) them upward through a tree. Now, all of them maintain
an error buffer, and the quasi-exception-style upward error propagation
is managed with a stand-in type, called `CavyError` (an error type that
should never be reported as such).

Other improvements:
* Command-line parsing has been factored out into more functions
* The `compile` API is very much simplified, and a lot of extraneous
  lifetimes have been removed.
* Some of the `resolve_annot` method has been written for the semantic
  analysis phase.

Shortcomings:
* `Arch`s were copied as an expedience, defeating the purpose of passing
  an immutable reference to `Session` in each compiler phase.
* Errors are actually not reported now; this must be restored.
* The repl is broken; apparently it is on the wrong compiler phase or
  something
* The `ErrorBuf` struct's `push` method returns a `CavyError` as an
  expedience; however, one must now be careful not to write
  `...ok_or(buf.push(...))?`, since this function evaluates its argument
  eagerly.