Newline annotations in tokenizing pass
The source store is now correctly updated with newline annotations by
the tokenizer. This made it necessary to pass a `&mut Session` to the
tokenizer along with a `SrcId`, which makes me wonder if the "notionally
mutable" parts of this struct should be broken out. Alternatively, we
could just recognize that the tokenizer doesn't need access to the whole
`Session`.

* Removed distinction between `SrcKind` and `SrcObject`; the source
  store now contains only `SrcObjects`, which are owned. Removing this
  arbitrary distinction should make it everywhere simpler to handle source
  code.
* Fixed an off-by-one error in the source-line-slicing code