Refactored into littler crates
There are now three crates contained in this repository: `cavy` (the
main library), `cavy-cli` (the command-line interface), and
`cavy-macros` (proc macros used by the library). This has a couple of
advantages: for one, it splits up the library and cli dependencies,
allowing us to track them separately. It also lets us build the library
alone, to export functions to alternative drivers of the compiler
written for scripting languages, or to target wasm and distribute the
whole compiler within a web application (if it's small enough!).