Added interners
`store.rs` now provides both `Store`s and `Interner`s. The former work
the same as they did before: you put in a stored value, and get back an
index with which you can look up the value later. An `Interner`, by
contrast, is "insert-only": you can insert a value and get back an
index, or check the index of a value. This addition made string
interning work properly, and the shadowed-variable error is now emitted.