CtxFmt trait to `impl Display` with aux Context
We'll often want to print things that can't be represented without some
additional information from the `Context`, e.g. interned symbols and
types. The "right" way to do this is with a wrapper type holding a
`&Context` that implements `Display`. This trait gives a method for
building these.

Also added some changes to the naming and location of the type interner
that should have been in the last commit.