Split `compile::compile` for circuit and target
The `compile` function has been split into two: one which returns a
circuit, and one which returns object code according the configuring
target. Both of these functions also now take a `SrcId` argument for the
entry point. This was a tiny change, but both "breaking" and atomic
enough to be its own commit.