Added SummaryRunner for simple analyses
Renamed `AnalysisRunner` to `DataflowRunner`; the new type of "runner"
is used for very simple analyses that don't really depend on the
control-flow graph topology, and could have been performed on the
AST (but appear here for the sake of consistency). The motivating
example is that of generating a call graph to detect mutual recursion
and ignore subgraphs not connected to `main` in codegen.

We may want both types of runner to implement a common trait, but for
now I've deferred this abstraction.