Start of call graph analysis
Started to work on call graph analysis. Got to producing a summary
analysis of all the call sites in each MIR graph. This is a good deal of
the way there: in principle, all we need to complete this analysis is a
cycle-detection routine.

The other thing that's left is really cleaning up the analysis module.
There's something quite off about how it's factored, that's making the
top-level function in `mod.rs` a busier than it should be.