Dirty miscellany: formatting, lowering
Mostly, major changes have been made to `cavy-macros`. We want to be
able to report errors with multiple `Spans`, and to this end have added
another `Diagnostic` method for "secondary" spans. This sort of works,
but has made the whole derive macro a mess, and caused a proliferation
of attributes. There's no way, currently, to put a help message on the
primary span. I should consider moving to a *single* `Vec<Span>`, which
is required to be nonempty, without treating the "main" span in any
special way. Then the principal error message doesn't have to be
attached to any span, which it isn't conceptually.

Added a few lines to `lowering.rs`.