Bugfix: error reporting with tabs
Previously, error messages would not line up correctly with source code
excerpts if they were written with tabs, rather than spaces. This has
been (mostly) fixed, but it *might* not work in all cases. We're just
counting the number of *leading* tabs, so an underline will not have the
right length under a span *containing* them. It's not clear to me how
such a span could possibly be formatted correctly, because we
don't (can't?) know tab stop length.

There will also be alignment errors if the source contains long unicode
characters, as in CJK.