Skeleton of optimization module and comptime
Added a sketch of an optimizations entry point. Each optimization will
just be a function that takes a mutable reference to the Mir. The start
of one such function has been given in `cavy-core/src/opt/comptime.rs`.
This will start out as a purely intraprocedural optimization, but it
might end up having to do something more complicated. Ultimately, it
might make sense for the Mir graphs to share structure, so that it's not
too expensive to recompile functions many times with different
parameters known at compile time.