Basic constant propagation / comptime evaluation
The `opt` module has a very simple interpreter in `comptime.rs`, which
for the moment can pretty much just NOT bools. These simple computations
are propagated through the MIR, and optimized-out statements are
replaced with a NOP.

Added some command-line arguments to set the final compiler phase to
post-optimization and to disable compile-time evaluation.