Bugfix: order truthy and falsy branches correctly
I accidentally put the `then` and `else` branches of an `if`
expression/statement in the wrong order in the MIR (the sensible order
is `else`, then `then`). This confusion came about from cryptically
naming the branches `direct` and `indirect`, which not only doesn't make
sense on the target architecture, it hides which one is which! This is a
good lesson in the importance of clear, straightforward naming.