Began change from mere `Local`s to `Place`s
(Again) following rustc's MIR, many locations where a simple `LocalId`
appeared before has been replaced with a `Place`, a location where some
data is stored, which is approximately a path of dot-operator-separated
names. In many cases this will not be much more expensive, because a
`Vec` doesn't allocate until its length is nonzero.

This is a first, extremely rough translation to the new `Place` system.
That means there are a lot of `clone`s (`Place` is not `Copy`), and many
spots where the root of the `Place` (a `LocalId`) is used without
change. This will continue to be a bit ugly for a little while yet.