Removed serde dependency
Panics are no longer logged in a file somewhere. For one, that's a
little bit intrusive. It's better not to make unexpected files if you
can help it. Second, serde was just making the banaries bigger and build
times longer for no real gain. The core library now has no external
dependencies.