Source and error reporting simplifications
A `SrcPoint` is now just a type alias for a `usize`, instead of carrying
around its line and column number. This means the scanner no longer has
to keep track of all these numbers. Instead, it just annotates the
locations of newlines (question: what happens if newlines are `\r\n`
newlines?). Line and column numbers are only computed at display time.
This has the side-effect of reducing the memory footprint, and making it
much easier to read the AST. Finally, single-character spans now have a
special-cased interval format!