Start of user-defined type resolution
Type aliases are now resolved in the MIR. The solution is quite a hack:
see the large comment above `lowering::typing::resolve_type`. Since both
the `MirBuilder` and `GraphBuilder` need to have access to this
resolution function, a great deal of data must be explicitly passed as
parameters--in slightly different ways, depending on the call site.
These two data structures also have to hold an extra table `udt_tys` to
translate `UdtId`s to `TyId`s after user-defined types are
fully-resolved as types.

All this is to say that this now works, but it feels like this part of
the lowering module is not factored in the natural, correct way. We'll
have to figure this out and hopefully fix it in a later cleanup pass.