Compile-time evaluation with tuples; Bugfix
Can now compute at compile time values from tuple fields. This works in
the basically naive way where the interpreter environment maintains
trees for each local. Since the type (and therefore size and layout) of
a local is known, we could do a one-off allocation instead of building
up these trees in a sort of ad-hoc way as values are moved into them.

Also, fixed a bug in `lowering.rs` that prevented the use of binop
expressions as tuple fields.