Lowering of struct literals
Struct literals are now lowered to tuples in the MIR. As usual, here are
the caveats of the implementation: we have to check at some point the
provided fields are exactly identical to those of the struct. To report
any errors, we have to report the symmetric difference of these sets of
names. In order to save an allocation or two, the two sides of the
symmetric difference are being performed in separate places: one in
`type_struct`, and one in `lower_into_struct`. This makes the logic
quite a bit more confusing, even with a signpost comment drawing
attention to this. If structs end up turning into a plate of spaghetti,
this could be one reason why.