Added `Translation` phase
In reality, our codegen is taking place in two phases: instead of a more
complex monolithic procedure, it consists of a lightweight translation
from the MIR, preserving the procedural structure, and a "zipping up" of
the procedures into an iterator over gates. These are now separated, and
the first of these is called `Translation`. This should make codegen
development and debugging easier.