Caching of type size and structural properties
The bit sizes and structural properties (i.e. linearity) of types are
now cached when they're interned. This is essentially a performance
optimization, but also a logic simplification for codegen: sizes won't
have to be computed recursively.

The `Context` now has methods `intern_ty` and `intern_symb`. This could
be simplified by abstracting over interners with an `Intern<I, V>`
trait, implemented by the `Interner` struct proper, as well as the new
`CachedTypeInterner` and the `Context`.