Bugfix: const prop through linear struct field
Test `const_prop_linear_struct_field` now passes. This previously failed
because, searching through a path during const propagaion,
`Value::follow` would attempt to access a nonexistent field with
`Value::slot`, which suffered an out-of-bounds panic. `slot` returns
`Option` and `follow` short-circuits if it reaches a missing field.

The simplest implementation results in `follow` and `follow_mut` having
different return types (`Option<&Value>` and `&mut Value`) than the
names might suggest. This might be a little confusing, so it might be
worth renaming them or changing the implementation in the future.