Delimited list parsing abstraction
The parser now has a new method, `delimited_list` that takes a function
and some configuration arguments, to produce a `Vec` of the type
returned by that function. To support this, all the delimiter lexemes
have been replaced with a `Lexeme::LDelim` and `Lexeme::RDelim`, and an
enum of delimiter kinds.