`OwnedBits` for Lir funcall args
Function arguments in the low-level intermediate representation are now
an `OwnedBits`. To accomplish this we've allocated for constants. There
are a few semi-ugly implementations with (in-principle) unnecessary
allocations, that are now marked with [PERF]. This will be the new
go-to, now that I'm less concerned with getting everything absolutely
optimal.