New function and user-defined type resolution
Previously, functions and types were simply looked up in the current
scope non-recursively. There's also a major fix here: the `with_table`
method on the `mir::GraphBuilder`, mirroring that on the `Parser`,
actually enters a Block scope when lowering it. So it's possible now to
call nested functions.

But there's a cost. The hideous type signature of
`lowering::typing::resolve_annot` has gotten even worse, since we need
not just the current table, but the whole `TableStore` to look things up
in outer scopes. This is clearly a mess in desperate need of
refactoring, once I have the freedom to make all this code beautiful and
elegant, rather than just "good enough."