Bugfix: field access on groups
`field_access` has been made its own BNF node. It replaces `primary` and
is now the sole parent of `primary`. This makes it possible to write
field accesses against groups, like `(a).bcd`. The catch is that numeric
literals are primary expressions, so there might be some parsing
ambiguity if floating-point numbers are ever added.