Bugfix: extra allocation for tuples
Previously, tuples allocated twice: once when the tuple's root was
inserted into the bindings, then again when the fields were populated.
We're now initializing the tuple's bindings with all zeros so this
doesn't happen. This section is still pretty inefficient, and like
others, will need to be fixed up when I have the time to do things
right. But we do get the right number of allocations now.