Restored LaTeX target
LaTeX backend now works with the new `Lir`. In order to support circuits
with an a priori unknown (but *hopefully* finite) number of bits, the
backend implementation has been changed to add new wires on-the-fly as
gates are inserted. This might be a little bit less efficient, and it's
probably a little bit uglier. As I'm approaching a deadline, code
quality is suffering a little bit, but it should be clearly marked.