Changed `Ext` statement to `Io` statement.
The `io` statement may now have both an input form and an output form.
There are currently problems with output of const values: this doesn't
really make sense. I think we want to insist that output const
values *not* be erased; that they must in fact live on the target
device. This means either re-inserting a statement in the mir (possibly
somewhat difficult), or doing some backwards control-flow analysis to
determine provenance (possible rather difficult).