Predecessor graph implementation from rustc
There's a bug--or not a bug, but a missing part of the implementation.
Constant propagation so far only propagates through the *first* basic
block. In order to propagate through later blocks, we need to unify the
state of the interpreter memory in multiple predecessor blocks.
Traditionally, this is done with phi functions in SSA, but I won't use
SSA. I do need a predecessor graph, though.

Question: will constant propagation be potentially O(n^2) in the number
of basic blocks, due to predecessor graph invalidation and re-computation?

The CFG now has a lazily-computed predecessor graph that's *directly*
taken from rustc. The implementation is changed only slightly. Because
I'm almost a little embarrassed about how similar this CFG is to rustc's
MIR, I've also added a comment to the README announcing that from the outset.