Swapped backends; set up basic data structures
I've now swapped out the more "static" backend using a low-level IR for
the more "dynamic" one, which is basically a CFG-level interpreter. Half
of this commit is merely shuffling types around to accommodate the
change. The other half is getting the ownership relations within the
`Interpreter` struct right. I think they seem reasonable enough now to
get to work.