Rewrote CLI target options
`latex_standalone` has been removed as a target; instead, there are
separate options for target settings including this one. It's not clear
if there's a way for clap to require one option conditional on the value
of *another* option, so I'm not sure if this setup can be "perfected" by
making e.g. the `--standalone` flag illegal if `--target qasm`.

Also, moved default values to the clap `.yml` specification.