General multiqubit gates
The general gate representation is now a struct containing a vector of
controls and a target "base gate". This is a little funny, because SWAP
is counted as a base gate, although it might decompose further into
single-qubit gates. So the representation might not be completely right
just yet, but it's getting more right.