Wrapped dataflow args into `DataflowCtx` struct
This should handle e.g. the precomputation of graph properties, making
them available to all dataflow analyses. The name is a little confusing,
though, because so far the word "context" has only been used in
connection with the `context::Context` struct, which is in fact a member
field of the `DataflowCtx`. Beauty is not a goal right now.