Point at infinity
Added a *single* lifetime (at most one, really) that extends into the
caller. This is actually a pretty severe restriction; all functions are
effectively constrained to look like:

```
fn f['a](x: &'a T, b: U, c: V['a]) -> &'a W
```

That's not ideal. But it is *much* easier to get a simple version of it
up and running, because we don't have to run a unification over the
references in the function signature. We would have to do so
either *before* lowering function bodies--and then figure out how to
thread the bound-lifetime ascriptions through the borrow checker for
each function, or compute them both within the function body and
at *each* call site. The latter doesn't sound right, and I'm not going
to do the former right now.

The lifetime constraints from function calls are now applied at the call
site.