Dataflow modifications and comments
The `bottom` method has been moved from `Lattice` to `DataflowAnalysis`.
This had to happen, unfortunately, because the "shape" of the `Lattice`
data can depend on essentially arbitrary types. I didn't want to add any
new generic bounds to this module.

I've also added some big comments on the relationship of middle-end
components with `rustc` for future readers who might be interested in
understanding the relationship between the compilers (particularly which
components are inspired by `rustc`, which are directly imitated in
simplified form, and which are unrelated), how this one was engineered,
and maybe how to read the `rustc` source.