Put old affine use and feedback analyses back in
Not only is the affine analysis pretty inefficient, I'm not sure it's
completely correct: it relies on every "double move" being propagated
all the way to the exit node (is it possible that there are could ever
be blocks that never reach it?).

I suspect that it should maybe be integrated with borrow checking. In
fact, I should probably do whatever rustc does, but there's no time for
a rewrite right now.

Similar commends on the old feedback analysis. Somehow these two aren't
doing quite the "right thing".