Finer check actions; lts dangle; all tests pass
The `{Shallow,Deep}{Read,Write}` actions described in the RFC have been
abstracted somewhat less, in order to make error messages more palatable.

Drop *never* counts toward liveness now. This is as if all references
were "may-dangle", as per the RFC. This is probably not quite right, but
it's a good place to start.