Static bit allocation
This is the biggest stage of the big refactoring. Bits are no longer
allocated dynamically: instead, we'll rely on optimization to ensure
that we don't swap too much or use too many bits.

This means the backend is tracking much less mutable state. Indeed, it
now has only *one* side-effect channel: the emitted gate stream. This
should be much easier to reason about, easier to program and less buggy.