Uncompute step now using recomputed lifetimes
There are edge cases that aren't quite right: in particular, *empty*
lifetimes are cleaned up at the *very end*--about as suboptimal as can
be. We must either optimize these out, or else ensure that their
destructors run immediately.

(Not only is this suboptimal, it's manifestly incorrect: we have references
outliving their referents, which you can see in the circuit diagrams.)

So, in fact, the times at which destructors execute are not *exactly*
the endpoints of lifetimes.