CircuitAssembler; arb phase gate; circuit tests
Refactored the `CircuitAssembler` struct in the dynamic backend. It's no
longer held in an `Rc<RefCell<_>>`; instead, there's a main `Circ`
struct (for lack of a better name), which can be borrowed to include
sinks (or not). This feels like a better way to factor the module, and
should simplify all the computation and gate assembly methods. It might
be less efficient due to constantly checking whether the sinks exist or
not; but it might be more efficient due to removing the smart pointers.
Who knows--simplicity is probably more important, anyway.

Also, finally added an arbitrary-phase gate, so we can demo all the "real"
algorithms.