In I/O, incomplete `Target` refactoring
Added In I/O instructions. In what should maybe be (part of a)
separate (series of) commit(s), the `Target`s are starting to be
refactored to allow the propagation of symbols through the end of
compilation. We're starting to do this by giving a `Context` to the
targets. I'm not really happy with that; it's surely not the "correctly"
factored solution.