Targets refactoring
Removed the `Context` fields from the `Qasm` and `Latex` targets--that
was not the right way to factor this. Now, many of the types in these
targets' modules implement `FmtWith` for a pair of the target and a
`Context`. This introduces less needless coupling, but it *looks*
complicated, and probably still isn't quite the right way to factor
this. But it's better.