Removed lifetime comparisons across linearization
Previously, *lots* of tests were failing because of attempted lifetime
comparisons between the lhs and rhs of statements like, `let x =
?true;`. These have been eliminated by adding a special case for
the `Linearization` unop when collecting constraints.

This feels like perhaps the less principled approach to making this
work. It might be necessary to consider a different one in the (near or
not-so-near) future.